Complete phases 1-2
